ENH merged test_crossval into test criterion
